Draw synapses as contact markers
Connectome viewers annotate synapses as discrete points rather than leaving them
implied by a connecting line, because where a contact sits along a process is
itself the finding: a cell targeted on its soma and one targeted on a distal
dendrite are wired differently even though the wiring diagram is identical.

Markers are placed with the same arc sampler the axon ribbon shader was verified
against, so a marker sits exactly on its ribbon rather than floating near it.
Coloured by receptor, sized sub-linearly in weight so one strong contact does not
hide its neighbours, and drawn as round sprites — a square point sprite reads as
a voxel, which is the impression this project exists not to give.

Named the contact accessor markerCount rather than count: THREE.Points already
declares one, and shadowing a base member to save four characters is how a subtle
clash gets introduced.

+++ apps/web/src/lib/similarity.ts
@@ -408,7 +408,13 @@ export function similarCells(prints: Fingerprints, slot: number, k: number): Sim
 
   const wanted = Math.max(1, Math.min(Math.floor(k), MAX_SIMILAR, Math.max(1, count - 1)));
   const bestSlots = new Int32Array(wanted);
-  const bestScores = new Float32Array(wanted);
+  // Float64, not Float32: the dot products accumulate in double, and rounding
+  // the kept scores to single would make two cells whose cosines differ in the
+  // ninth decimal compare equal — near 1, where the interesting matches are, a
+  // float32 step is 6e-8 wide. Comparing a double against a rounded-down stored
+  // score also lets a strictly worse later cell displace a better earlier one,
+  // which is the tie rule below running backwards.
+  const bestScores = new Float64Array(wanted);
   let length = 0;
 
   const query = slot * dim;
